requirements.txt
pandas
tensorboard
pyyaml
munch
tabulate
colorama